requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.112.0
h11==0.14.0
idna==3.7
iso8601==1.1.0
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.21.5
typing_extensions==4.12.2
uvicorn==0.30.5
